pyproject.toml
[project]
name = "SideBot"
version = "0.0.1"
authors = [
    { name="nythepegasus", email="me@nythepegas.us" },
]
description = "SideBot is a multipurpose bot for the SideStore community."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.3.2",
    "psycopg2==2.9.9",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

[project.optional-dependencies]
dev = [
    "build",
    "pylint",
    "mypy",
    "ruff",
    "types-psycopg2",
]
